requirements.txt
Appium_Python_Client==4.5.1
autogen_agentchat==0.4.8.2
autogen_core==0.4.8.2
autogen_ext==0.4.8.2
colorlog==6.9.0
Jinja2==3.1.3
PyYAML==6.0.2
PyYAML==6.0.2
selenium==4.29.0
tiktoken==0.8.0
